[refactor][move] v2.0.0 FRONTEND 구조화 변경 AdminLayout 경로 변경으로 변경된 파일 REF : #11
